Try fix tests
